Affine Cipher
Added new cipher, Affine that performs affine cryptography transformations (of the form ax + b). The affine cipher needs 'a' and 26 to be coprime and will throw an exception otherwise.
Added utils package object to hold utility functions and classes. Currently contains euclid and modInv methods for use in the affine cipher